+++ uefi/src/mem/memory_map/impl_.rs
@@ -432,6 +432,7 @@ mod tests {
     const BASE_MMAP_UNSORTED: [MemoryDescriptor; 3] = [
         MemoryDescriptor {
             ty: MemoryType::CONVENTIONAL,
+            padding0: 0,
             phys_start: 0x3000,
             virt_start: 0x3000,
             page_count: 1,
@@ -439,6 +440,7 @@ mod tests {
         },
         MemoryDescriptor {
             ty: MemoryType::CONVENTIONAL,
+            padding0: 0,
             phys_start: 0x2000,
             virt_start: 0x2000,
             page_count: 1,
@@ -446,6 +448,7 @@ mod tests {
         },
         MemoryDescriptor {
             ty: MemoryType::CONVENTIONAL,
+            padding0: 0,
             phys_start: 0x1000,
             virt_start: 0x1000,
             page_count: 1,